Fix SMALLINT overflow for runtime_minutes
IMDb contains titles with runtimeMinutes > 32767 (e.g. 37440), which
exceeds SMALLINT's max. Changed runtime_minutes to INTEGER in both the
staging table DDL and the schema reference.

+++ etl/main.py
@@ -150,7 +150,7 @@ def create_staging_tables(conn):
                 title_type VARCHAR(20) NOT NULL,
                 primary_title TEXT NOT NULL,
                 start_year SMALLINT,
-                runtime_minutes SMALLINT,
+                runtime_minutes INTEGER,
                 genres TEXT
             )
         """)
